refactor: move the HTTP semantics helpers out of the package controller
Compression pushed package-controller.go from 260 to 367 lines, and the
growth landed in the worst place: 130 lines of pure functions sat between
createPackage and getPackage, so the handlers could no longer be read in
sequence.

Content negotiation moves to content-negotiation.go and the conditional
request helpers to conditional-requests.go, each with its own test file.
The split follows the subjects rather than "not a handler": negotiation
comes from #47, conditional requests from #48, and they evolve apart.

The helpers stay unexported in package api. A subdirectory would make them
a package boundary and force six symbols public for a single caller, and
templates.go already shows the package holding files named after what they
do rather than after a controller.

package-controller.go is now 234 lines, below its size before compression.
No behaviour change: the moved tests are the ones that prove it.
