+++ api/package-controller_test.go
@@ -11,7 +11,6 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
@@ -281,121 +280,4 @@ func TestGetSlice_Valid(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Equal(t, "mypkg", slice.Name)
 	assert.Equal(t, "1.2.3", slice.Version)
-}
-
-func TestCacheControlHeader(t *testing.T) {
-	tests := []struct {
-		version  string
-		expected string
-	}{
-		{"1.0.0", "public, max-age=31536000, immutable"},
-		{"1.0.1", "public, max-age=31536000, immutable"},
-		{"1.0", "public, max-age=300"},
-		{"1", "public, max-age=300"},
-		{"1.0.0-rc.1", "public, max-age=300"},
-		{"1.0.0+build.1", "public, max-age=31536000, immutable"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.version, func(t *testing.T) {
-			assert.Equal(t, tt.expected, cacheControlHeader(tt.version))
-		})
-	}
-}
-
-func TestETagMatches(t *testing.T) {
-	tests := []struct {
-		name        string
-		ifNoneMatch string
-		etag        string
-		expected    bool
-	}{
-		{"exact match", `"abc"`, `"abc"`, true},
-		{"weak request ETag", `W/"abc"`, `"abc"`, true},
-		{"weak response ETag", `"abc"`, `W/"abc"`, true},
-		{"match in second position", `"stale", W/"abc"`, `"abc"`, true},
-		{"wildcard", `*`, `"abc"`, true},
-		{"stale value", `"stale"`, `"abc"`, false},
-		{"empty header", ``, `"abc"`, false},
-		{"empty ETag", `*`, ``, false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.expected, etagMatches(tt.ifNoneMatch, tt.etag))
-		})
-	}
-}
-
-func TestParseAcceptEncoding(t *testing.T) {
-	br := content.EncodingBrotli
-	gzip := content.EncodingGzip
-	tests := []struct {
-		header    string
-		encodings []content.Encoding
-		identity  bool
-	}{
-		{"", nil, true},
-		{"br", []content.Encoding{br}, true},
-		{"gzip", []content.Encoding{gzip}, true},
-		{"br, gzip", []content.Encoding{br, gzip}, true},
-		{"br;q=0.5, gzip;q=0.9", []content.Encoding{gzip, br}, true},
-		{"gzip;q=0", nil, true},
-		{"*", []content.Encoding{br, gzip}, true},
-		{"deflate, zstd", nil, true},
-		{"br;q=nope, gzip", []content.Encoding{gzip}, true},
-		// A named entry wins over "*" whatever the order, so an explicit refusal
-		// is not resurrected by the wildcard (RFC 9110 section 12.5.3, rule 3).
-		{"gzip;q=0, *;q=1", []content.Encoding{br}, true},
-		{"*;q=1, gzip;q=0", []content.Encoding{br}, true},
-		{"br;q=0, *", []content.Encoding{gzip}, true},
-		{"*;q=0.5, br;q=0.1", []content.Encoding{gzip, br}, true},
-		{"br;q=0.1, *;q=0.5", []content.Encoding{gzip, br}, true},
-		// The unencoded representation is refused only by an entry that names it,
-		// or by a wildcard no identity entry overrides.
-		{"identity;q=0", nil, false},
-		{"*;q=0", nil, false},
-		{"*;q=0, identity", nil, true},
-		{"identity;q=0, br", []content.Encoding{br}, false},
-		{"gzip;q=1.0, identity;q=0.5, *;q=0", []content.Encoding{gzip}, true},
-	}
-	for _, tt := range tests {
-		accepted := parseAcceptEncoding(tt.header)
-		assert.Equal(t, tt.encodings, accepted.encodings, tt.header)
-		assert.Equal(t, tt.identity, accepted.identity, tt.header)
-	}
-}
-
-func TestNotModified(t *testing.T) {
-	lastModified := time.Date(2026, time.August, 21, 12, 30, 45, 0, time.UTC)
-	tests := []struct {
-		name            string
-		ifNoneMatch     string
-		ifModifiedSince string
-		lastModified    time.Time
-		etag            string
-		expected        bool
-	}{
-		{"matching If-None-Match", `"abc"`, "", lastModified, `"abc"`, true},
-		{"stale If-None-Match takes precedence", `"stale"`, lastModified.Format(http.TimeFormat), lastModified, `"abc"`, false},
-		{"If-Modified-Since exact match", "", lastModified.Format(http.TimeFormat), lastModified, `"abc"`, true},
-		{"If-Modified-Since older", "", lastModified.Add(-time.Second).Format(http.TimeFormat), lastModified, `"abc"`, false},
-		{"malformed date", "", "not-a-date", lastModified, `"abc"`, false},
-		{"zero last modified", "", lastModified.Format(http.TimeFormat), time.Time{}, `"abc"`, false},
-		{"sub-second last modified", "", lastModified.Format(http.TimeFormat), lastModified.Add(987 * time.Millisecond), `"abc"`, true},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			r := httptest.NewRequest(http.MethodGet, "/", nil)
-			if tt.ifNoneMatch != "" {
-				r.Header.Set("If-None-Match", tt.ifNoneMatch)
-			}
-			if tt.ifModifiedSince != "" {
-				r.Header.Set("If-Modified-Since", tt.ifModifiedSince)
-			}
-
-			assert.Equal(t, tt.expected, notModified(r, tt.etag, tt.lastModified))
-		})
-	}
 }
